backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
python-multipart==0.0.9

onnxruntime==1.17.1
insightface==0.7.3
faiss-cpu==1.8.0